+++ packages/langium/src/default-module.ts
@@ -24,7 +24,8 @@ import { DefaultValueConverter } from './parser/value-converter';
 import { DefaultLinker } from './references/linker';
 import { DefaultNameProvider } from './references/naming';
 import { DefaultReferences } from './references/references';
-import { DefaultScopeComputation, DefaultScopeProvider } from './references/scope';
+import { DefaultScopeComputation } from './references/scope-computation';
+import { DefaultScopeProvider } from './references/scope-provider';
 import { DefaultJsonSerializer } from './serializer/json-serializer';
 import { DefaultServiceRegistry } from './service-registry';
 import { LangiumDefaultServices, LangiumDefaultSharedServices, LangiumServices, LangiumSharedServices } from './services';
