+++ packages/langium/src/index.ts
@@ -25,7 +25,8 @@ export * from './parser/token-builder';
 export * from './parser/value-converter';
 export * from './references/linker';
 export * from './references/naming';
-export * from './references/scope';
+export * from './references/scope-computation';
+export * from './references/scope-provider';
 export * from './serializer/json-serializer';
 export * from './service-registry';
 export * from './services';
