+++ packages/langium/src/services.ts
@@ -29,7 +29,8 @@ import type { ValueConverter } from './parser/value-converter';
 import type { Linker } from './references/linker';
 import type { NameProvider } from './references/naming';
 import type { References } from './references/references';
-import type { ScopeComputation, ScopeProvider } from './references/scope';
+import type { ScopeComputation } from './references/scope-computation';
+import type { ScopeProvider } from './references/scope-provider';
 import type { JsonSerializer } from './serializer/json-serializer';
 import type { ServiceRegistry } from './service-registry';
 import type { AstReflection } from './syntax-tree';
